+++ SSD1306Brzo.h
@@ -84,7 +84,7 @@ class SSD1306Brzo : public OLEDDisplay {
        // If the minBoundY wasn't updated
        // we can savely assume that buffer_back[pos] == buffer[pos]
        // holdes true for all values of pos
-       if (minBoundY == ~0) return;
+       if (minBoundY == (uint8_t) ~0) return;
 
        sendCommand(COLUMNADDR);
        sendCommand(minBoundX);
